aws/lambdas/percentile/requirements.txt
colorama==0.4.6
et_xmlfile==2.0.0
iniconfig==2.1.0
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.15.3
six==1.17.0
tzdata==2025.2
PyJWT==2.8.0
cryptography==41.0.8
requests==2.31.0